requirements.txt
requests
bs4
simplejson
git+https://github.com/Ashwin-Surana/scrapkit
django
djangorestframework
mongoengine